perf: improve startup with async imports
