+++ src/start.ts
@@ -1,5 +1,4 @@
 import 'dotenv/config';
-import './env-vars-check';
 import { start } from './server';
 
 start();
