+++ src/utils/env.ts
@@ -190,4 +190,11 @@ export const ENV = {
   get AUTH_USER_SESSION_VARIABLE_FIELDS(): Record<string, string> {
     return this.AUTH_USER_SESSION_VARIABLE_FIELDS;
   },
+  // * See ../server.ts
+  // get AUTH_SKIP_INIT() {
+  //   return castBooleanEnv('AUTH_SKIP_INIT', false);
+  // },
+  // get AUTH_SKIP_SERVE() {
+  //   return castBooleanEnv('AUTH_SKIP_SERVE', false);
+  // },
 };
